sanctum-server/requirements.txt
pinecone-client
openai
python-dotenv
flask
scipy
flask_cors
grpcio